fix: 분리된 타입 파일 정리 및 홈 헤더의 셀렉트 박스 수정, 상세페이지 Routes 경로수정
